+++ ifs-web-service/ifs-competition-mgt-service/src/test/java/org/innovateuk/ifs/management/controller/CompetitionManagementPanelInviteAssessorsOverviewControllerTest.java
@@ -164,7 +164,7 @@ public class CompetitionManagementPanelInviteAssessorsOverviewControllerTest ext
             assertEquals(inviteOverviewResource.getBusinessType(), overviewAssessorRowViewModel.getBusinessType());
             assertEquals(inviteOverviewResource.getStatus(), overviewAssessorRowViewModel.getStatus());
             assertEquals(inviteOverviewResource.getDetails(), overviewAssessorRowViewModel.getDetails());
-            assertEquals(inviteOverviewResource.getInviteId(), overviewAssessorRowViewModel.getInviteId());
+            assertEquals((long) inviteOverviewResource.getInviteId(), overviewAssessorRowViewModel.getInviteId());
         });
     }
